rename : 파일이름에 sample 제거
